Add a build config file for the Turtle
This builds the js and d.ts files in the 'public' folder for consumption by index.html and the monaco editor.

Building / working inside this directory instead of the 'src' because the bigger project feels that src/turtle/sourceFile.ts and public/turtle/builtFiles.js/d.ts have duplicate definitions.